version control the object-cache dropin
There are some strange bugs happening pulling object cache via composer.
This has to do with how the dropin installer is interacting with the
git version controlled package. When the installer pulls out the
object-cache.php file, it makes the git repo have 'uncommitted changes'
and composer is failing. This in turn seems to be causing other
problems as well. So for now, rather than pull object cache through
composer we'll just version control the drop in. In the future, I'm
hopeful this will change.